+++ packages/bluesky/lib/bluesky.dart
@@ -347,6 +347,7 @@ export 'package:bluesky/src/services/entities/labeler_service_record.dart';
 export 'package:bluesky/src/services/entities/labelers_pref.dart';
 export 'package:bluesky/src/services/entities/labeler_pref_item.dart';
 export 'package:bluesky/src/services/entities/feed_interaction.dart';
+export 'package:bluesky/src/services/entities/suggestions_skeleton.dart';
 export 'package:bluesky/src/services/extensions/grouped_notification.dart';
 export 'package:bluesky/src/services/extensions/at_uri.dart';
 export 'package:bluesky/src/services/extensions/blob.dart';