+++ packages/bluesky/test/src/services/actor_service_test.dart
@@ -39,7 +39,6 @@ void main() {
   testActor<ActorsTypeahead>(
     (m, s) => s.searchActorsTypeahead(
       term: m.query,
-      viewer: m.did,
       limit: m.limit,
     ),
     id: appBskyActorSearchActorsTypeahead,