+++ packages/bluesky/test/src/services/unspecced_service_test.dart
@@ -7,6 +7,7 @@ import 'package:bluesky/src/ids.g.dart';
 import 'package:bluesky/src/services/entities/feed_generators.dart';
 import 'package:bluesky/src/services/entities/skeleton_actors_by_query.dart';
 import 'package:bluesky/src/services/entities/skeleton_posts_by_query.dart';
+import 'package:bluesky/src/services/entities/suggestions_skeleton.dart';
 import 'package:bluesky/src/services/entities/tagged_suggestions.dart';
 import 'suite/service_suite.dart';
 
@@ -32,5 +33,14 @@ void main() {
   testUnspecced<TaggedSuggestions>(
     (m, s) => s.getTaggedSuggestions(),
     id: appBskyUnspeccedGetTaggedSuggestions,
+  );
+
+  testUnspecced<SuggestionsSkeleton>(
+    (m, s) => s.getSuggestionsSkeleton(
+      viewer: m.did,
+      limit: m.limit,
+      cursor: m.cursor,
+    ),
+    id: appBskyUnspeccedGetSuggestionsSkeleton,
   );
 }